requirements.txt
# requirements.txt
pillow
selenium
webdriver-manager
tkinter-tooltip
requests
beautifulsoup4
lxml
pytesseract
opencv-python-headless
pywebview~=5.4
jinja2